잭슨 룰 설명

+++ src/main/java/com/sbs/basic1/boundedContext/member/entity/Member.java
@@ -1,13 +1,12 @@
 package com.sbs.basic1.boundedContext.member.entity;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @AllArgsConstructor
 @Data
 public class Member {
-  private static int lastId;
+  private static long lastId;
   private final long id;
   private final String username;
   private final String password;
